+++ coordinator/app/src/main/kotlin/net/consensys/zkevm/coordinator/app/CoordinatorApp.kt
@@ -8,9 +8,15 @@ import io.vertx.sqlclient.SqlClient
 import linea.coordinator.config.v2.CoordinatorConfig
 import linea.coordinator.config.v2.DatabaseConfig
 import linea.persistence.DisabledForcedTransactionsDao
+import linea.persistence.conflation.AggregationsRepositoryImpl
 import linea.persistence.conflation.BatchesPostgresDao
+import linea.persistence.conflation.BlobsPostgresDao
+import linea.persistence.conflation.BlobsRepositoryImpl
+import linea.persistence.conflation.PostgresAggregationsDao
 import linea.persistence.conflation.PostgresBatchesRepository
 import linea.persistence.conflation.RetryingBatchesPostgresDao
+import linea.persistence.conflation.RetryingBlobsPostgresDao
+import linea.persistence.conflation.RetryingPostgresAggregationsDao
 import linea.persistence.ftx.PostgresForcedTransactionsDao
 import linea.persistence.ftx.RetryingPostgresForcedTransactionsDao
 import net.consensys.linea.async.toSafeFuture
@@ -21,12 +27,6 @@ import net.consensys.linea.vertx.loadVertxConfig
 import net.consensys.zkevm.coordinator.api.Api
 import net.consensys.zkevm.coordinator.app.conflationbacktesting.ConflationBacktestingService
 import net.consensys.zkevm.fileio.DirectoryCleaner
-import net.consensys.zkevm.persistence.dao.aggregation.AggregationsRepositoryImpl
-import net.consensys.zkevm.persistence.dao.aggregation.PostgresAggregationsDao
-import net.consensys.zkevm.persistence.dao.aggregation.RetryingPostgresAggregationsDao
-import net.consensys.zkevm.persistence.dao.blob.BlobsPostgresDao
-import net.consensys.zkevm.persistence.dao.blob.BlobsRepositoryImpl
-import net.consensys.zkevm.persistence.dao.blob.RetryingBlobsPostgresDao
 import net.consensys.zkevm.persistence.db.Db
 import net.consensys.zkevm.persistence.db.PersistenceRetryer
 import org.apache.logging.log4j.Level
