+++ coordinator/persistence/src/integrationTest/kotlin/linea/persistence/conflation/BatchesPostgresDaoTest.kt
@@ -11,7 +11,6 @@ import net.consensys.FakeFixedClock
 import net.consensys.linea.async.get
 import net.consensys.zkevm.persistence.db.DbHelper
 import net.consensys.zkevm.persistence.db.test.CleanDbTestSuiteParallel
-import net.consensys.zkevm.persistence.db.test.DbQueries
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
