+++ coordinator/persistence/src/integrationTest/kotlin/linea/persistence/test/DbSchemaUpdatesIntTest.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.persistence.db.test
+package linea.persistence.test
 
 import io.vertx.core.Vertx
 import io.vertx.junit5.VertxExtension
@@ -8,6 +8,7 @@ import linea.kotlin.encodeHex
 import net.consensys.linea.async.get
 import net.consensys.zkevm.persistence.db.Db
 import net.consensys.zkevm.persistence.db.DbHelper
+import net.consensys.zkevm.persistence.db.test.DbQueries
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
