+++ coordinator/persistence/src/main/kotlin/linea/persistence/conflation/PostgresAggregationsDao.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.persistence.dao.aggregation
+package linea.persistence.conflation
 
 import io.vertx.core.Future
 import io.vertx.pgclient.PgException
