+++ coordinator/persistence/src/main/kotlin/linea/persistence/conflation/RecordsCleanupFinalizationHandler.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.persistence.dao.aggregation
+package linea.persistence.conflation
 
 import linea.finalization.FinalizationHandler
 import linea.finalization.FinalizationMonitor
