+++ coordinator/persistence/src/test/kotlin/linea/persistence/conflation/RetryingBlobsPostgresDaoTest.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.persistence.dao.blob
+package linea.persistence.conflation
 
 import io.vertx.core.Vertx
 import io.vertx.junit5.VertxExtension
